refactor: Move env-profile getters from constants to environment module
